partially there

+++ include/block_models.h
@@ -31,7 +31,7 @@ typedef struct {
     uint id;
     char* name;
     int transparent;
-    float face_atlas_coords[6][2];
+    int face_atlas_coords[6][2];
 } block_type;
 
 extern block_type TYPES[];
